+++ dst/app/routes/router.js
@@ -8,6 +8,7 @@ const express = require("express");
 const health_1 = require("./health");
 const _ah_1 = require("./_ah");
 const accountTransactions_1 = require("./accountTransactions");
+const cron_1 = require("./cron");
 const permits_1 = require("./permits");
 const authentication_1 = require("../middlewares/authentication");
 const router = express.Router();
@@ -21,6 +22,7 @@ router.get('', (__, res) => {
     res.send('hello!');
 });
 router.use('/_ah', _ah_1.ahRouter);
+router.use('/cron', cron_1.cronRouter);
 router.use('/health', health_1.healthRouter);
 // 認証
 router.use(authentication_1.authentication);