+++ src/app/middlewares/authentication.ts
@@ -1,7 +1,7 @@
 /**
  * oauthミドルウェア
  */
-import { chevre } from '@cinerino/domain';
+import { chevre } from '@chevre/domain';
 import { cognitoAuth } from '@motionpicture/express-middleware';
 import * as createDebug from 'debug';