+++ src/app/middlewares/notFoundHandler.spec.ts
@@ -2,7 +2,7 @@
 /**
  * not foundハンドラーミドルウェアテスト
  */
-import { chevre } from '@cinerino/domain';
+import { chevre } from '@chevre/domain';
 import * as assert from 'assert';
 import * as nock from 'nock';
 import * as sinon from 'sinon';