+++ src/app/middlewares/notFoundHandler.ts
@@ -1,7 +1,7 @@
 /**
  * 404ハンドラーミドルウェア
  */
-import { chevre } from '@cinerino/domain';
+import { chevre } from '@chevre/domain';
 import { NextFunction, Request, Response } from 'express';
 
 export function notFoundHandler(req: Request, __: Response, next: NextFunction) {